Refactor: string folder structure
