+++ src/components/TagColorMenu/components/ColorItem/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import { findColorClass } from "src/services/color";
-import { uppercaseFirst } from "src/services/string/modify";
+import { uppercaseFirst } from "src/services/string/utils";
 import "./styles.css";
 
 interface Props {
