--- src/services/string/modify/index.ts
@@ -1,11 +1,0 @@
-export const stripPound = (tag: string) => {
-	return tag.replace("#", "");
-};
-
-export const uppercaseFirst = (input: string) => {
-	return input.charAt(0).toUpperCase() + input.slice(1);
-};
-
-export const addPound = (input: string) => {
-	return `#${input}`;
-};
